+++ sdk/src/serum/types.ts
@@ -5,9 +5,13 @@ export type SerumMarketSubscriberConfig = {
 	connection: Connection;
 	programId: PublicKey;
 	marketAddress: PublicKey;
-	accountSubscription: {
-		// enables use to add web sockets in the future
-		type: 'polling';
-		accountLoader: BulkAccountLoader;
-	};
+	accountSubscription:
+		| {
+				// enables use to add web sockets in the future
+				type: 'polling';
+				accountLoader: BulkAccountLoader;
+		  }
+		| {
+				type: 'websocket';
+		  };
 };