+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - program: max number of subaccounts to 3000
+- program: add pyth1M/pyth1K as oracle scales
 
 ### Fixes